refactor(frontend): split bronnen.tsx into single-responsibility files (SPEC-PORTAL-SOURCES-RENAME-001 Phase 3)
bronnen.tsx (657 lines, god-component) replaced by:

  sources.tsx           151 lines  route shell — data fetching + layout only
  -sources-hooks.ts     173 lines  useSourceSync / useSourceDelete /
                                   useSourceRename / useSourceReauth /
                                   useSyncAllConnectors
  -sources-row.tsx      281 lines  SourceRow component (still contains the
                                   inline rename form; Phase 4 swaps to
                                   InlineEdit and brings this under 200)
  -sources-content.tsx  128 lines  Drill-down split into ConnectorContent
                                   + UploadContent components
  -sources-actionbar.tsx 72 lines  Top-of-tab "Open in editor" /
                                   "Synchroniseer alles" / "Bron toevoegen"
  -sources-helpers.tsx   86 lines  SourceIcon, mapSourceStatus, StatusBadge
  -sources-types.ts      48 lines  Source, SourcesResponse, ContentResponse,
                                   PageIndexEntry

Plus file path + URL renames so the SPEC's Phase 3.6 only needs the
legacy /bronnen → /sources redirect (TanStack file routing makes
sources.tsx own /sources today; route shell + all in-app navigation
already point to /sources).

Pencil-icon disambiguation and InlineEdit swap arrive in Phase 4.
Paraglide migration of the ~25 hardcoded NL strings (still present in
the split components 1:1 with the pre-split file) arrives in Phase 5.

Acceptance criteria met / deferred:
  AC-3 sources.tsx ≤ 200 ✅  (151)
  AC-4 -sources-row.tsx ≤ 200  deferred to Phase 4 (281 — InlineEdit
       removes the 60-line hand-rolled rename form)
  AC-5 -sources-content.tsx ≤ 120  marginal (128, 8 over; split-by-kind
       structure is the goal, not the line count)

Verified:
  tsc --noEmit (with --ignoreDeprecations 6.0): exit 0
  vitest run src/routes/app/knowledge: 38/38 pass
  routeTree.gen.ts regenerated: /sources route registered
