+++ klai-portal/frontend/src/routes/app/knowledge/$kbSlug_.add-source._components/FileUploadForm.tsx
@@ -209,7 +209,7 @@ export function FileUploadForm({ kbSlug, onBack }: FileUploadFormProps) {
         setAllDone(true)
         const t = setTimeout(() => {
           void navigate({
-            to: '/app/knowledge/$kbSlug/bronnen',
+            to: '/app/knowledge/$kbSlug/sources',
             params: { kbSlug },
           })
         }, 1500)
